feat: implement routing api

+++ Makefile
@@ -38,7 +38,6 @@ help:
 	@echo "  build                             Build"
 	@echo "  generate                          Generate"
 	@echo "  generate/client                   Generate client pkg"
-	@echo "  generate/ors_client		   Generate client pkg"
 	@echo "  setup                             Install dependencies"
 	@echo "  setup/macos                       Install dependencies for macOS"
 	@echo "  setup/ci                          Install dependencies for CI"
@@ -106,11 +105,6 @@ generate:
 generate/client: generate
 	@echo "Generating client..."
 	@./scripts/openapi-generator.sh client docs/swagger.yaml pkg/client
-
-.PHONY: generate/ors
-generate/ors: 
-	@echo "Generating ors api client..."
-	@./scripts/openapi-generator.sh ors pkg/ors_api/api-docs.json pkg/ors_api
 
 .PHONY: setup
 setup:
@@ -203,9 +197,11 @@ run/docker: run/docker/prepare
 infra/up:
 	@echo "Running infra..."
 	mkdir -p .docker/infra/ors/{config,elevation_cache,files,graphs,logs}
+	mkdir -p .docker/infra/valhalla/custom_files
 	chown -R $(USER_ID) .docker/infra/ors
 	yq e -i '.services."ors-app".user = env(USER_ID)' .docker/docker-compose.infra.yaml
-	wget https://download.geofabrik.de/europe/germany/schleswig-holstein-latest.osm.pbf -O .docker/infra/ors/files/sh.osm.pbf
+	test -f .docker/infra/ors/files/sh.osm.pbf || wget https://download.geofabrik.de/europe/germany/schleswig-holstein-latest.osm.pbf -O .docker/infra/ors/files/sh.osm.pbf
+	test -f .docker/infra/valhalla/custom_files/sh.osm.pbf || wget https://download.geofabrik.de/europe/germany/schleswig-holstein-latest.osm.pbf -O .docker/infra/valhalla/custom_files/sh.osm.pbf
 
 	docker compose -f .docker/docker-compose.infra.yaml up -d
